Fixed device deconstructors, fixed file mkdir on linux.
Former-commit-id: 49f08d87622127bc7c3b1fc86ee4af5697d7ac18
Former-commit-id: 7e93b3b8356291175168522cb74866c05ee8290b

+++ Sources/FlounderEngine/Files/Csv/FileCsv.hpp
@@ -34,7 +34,7 @@ namespace Flounder
 
 		void Save() override;
 
-		unsigned int GetRowCount() const { return m_rows->size(); }
+		size_t GetRowCount() const { return m_rows->size(); }
 
 		RowCsv GetRow(const unsigned int &index);
